pyproject.toml
[tool.poetry]
name="Tools Observatory"
version="0.1.0"
description="A collections of small tools to solve specific problems."
authors=["Gurupratap Matharu <gurupratap.matharu@gmail.com>"]
readme="README.md"

[tool.poetry.dependencies]
python="^3.11"
factory-boy="^3.3.0"
python-dotenv="^1.0.1"
django-extensions="^3.2.3"
django-allauth={extras=["socialaccount"], version="^65.11.2"}
django="^5.2.6"
pillow="^11.3.0"
django-taggit="^6.1.0"
markdown="^3.9"


[tool.poetry.group.dev.dependencies]
ruff="^0.4.6"
djlint="^1.34.1"
ipython="^8.24.0"
django-debug-toolbar="^6.0.0"

[build-system]
requires=["poetry-core"]
build-backend="poetry.core.masonry.api"

[tool.ruff]
line-length=119
exclude=["*migrations*", "static"]

[tool.ruff.lint.isort]
combine-as-imports=true
force-single-line=false
known-first-party=["blog"]
known-third-party=["rest_framework"]
lines-after-imports=2
section-order=["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
"django"=["django"]

[tool.djlint]
blank_line_before_tag="load,extends"
blank_line_after_tag="load,extends"
close_void_tags=true
custom_blocks="toc,example"
custom_html="mjml"
exclude=".venv,venv,.tox,.eggs,.git"
extend_exclude=".custom"
format_attribute_template_tags=true
format_css=true
format_js=true
ignore_blocks="raw,example"
indent=2
max_attribute_length=120
max_line_length=120
profile="django"
# preserve_blank_lines=true
use_gitignore=true

[tool.djlint.css]
indent_size=5

[tool.djlint.js]
indent_size=5

[tool.coverage.run]
branch=true
omit=[
    "*/migrations/*",
    "*/tests/*",
    "/usr/*",
    "main/asgi.py",
    "main/wsgi.py",
    "main/settings.py",
    "manage.py",
    "*/urls.py",
    "*/__init__.py",
]

[tool.coverage.report]
show_missing=true
skip_covered=false

[tool.coverage.html]
show_contexts=true
